Do not add bus parameter to EBS after 1.21.1
This also moves the client mod events to the client entrypoint for logical separation when after 1.20.5
